# [Pytoch][Vulkan] Create context for layernorm