--- a/aten/src/ATen/native/vulkan/ops/Layernorm.cpp
+++ b/aten/src/ATen/native/vulkan/ops/Layernorm.cpp
@@ -1,3 +1,9 @@
+#include <ATen/native/vulkan/ops/Layernorm.h>
+#include <ATen/native/vulkan/ops/Utils.h>
+
+#include <ATen/Context.h>
+#include <c10/util/irange.h>
+
 #include <ATen/native/vulkan/ops/Common.h>
 #include <torch/library.h>
 
@@ -11,9 +17,66 @@ namespace at {
 namespace native {
 namespace vulkan {
 namespace ops {
-namespace {
 
-using namespace api::utils;
+LayernormPackedContext::LayernormPackedContext(
+    const c10::optional<Tensor>& weight,
+    const c10::optional<Tensor>& bias,
+    double eps)
+    : unpacked_{c10::AnyType::get()} {
+  packed_.reserve(ListArgs::kNumArgs);
+
+  TORCH_CHECK(weight, "Weight must be provided!");
+  packed_.emplace_back(weight->vulkan());
+  TORCH_CHECK(bias, "Bias must be provided!");
+  packed_.emplace_back(bias->vulkan());
+  packed_.emplace_back(eps);
+
+  if (!at::globalContext().releaseWeightsWhenPrepacking()) {
+    unpacked_.reserve(ListArgs::kNumArgs);
+    unpacked_.emplace_back(weight);
+    unpacked_.emplace_back(bias);
+    unpacked_.emplace_back(eps);
+  }
+}
+
+LayernormPackedContext LayernormPackedContext::pack(
+    c10::impl::GenericList unpacked) {
+  return LayernormPackedContext(
+      get_optional_tensor(unpacked, ListArgs::kWeight),
+      get_optional_tensor(unpacked, ListArgs::kBias),
+      unpacked.get(ListArgs::kEps).toDouble());
+}
+
+c10::intrusive_ptr<LayernormPackedContext> create_layernorm_context(
+    c10::optional<Tensor>&& weight,
+    c10::optional<Tensor>&& bias,
+    double eps) {
+  return c10::make_intrusive<LayernormPackedContext>(
+      LayernormPackedContext(weight, bias, eps));
+}
+
+Tensor run_layernorm_context(
+    const Tensor& input_arg,
+    IntArrayRef normalized_shape,
+    const c10::intrusive_ptr<LayernormPackedContext>& layernorm_context) {
+  const Tensor input = input_arg.is_vulkan() ? input_arg : input_arg.vulkan();
+
+  const c10::optional<Tensor>& weight_opt =
+      layernorm_context->get_val(LayernormPackedContext::ListArgs::kWeight)
+          .toTensor();
+  const c10::optional<Tensor>& bias_opt =
+      layernorm_context->get_val(LayernormPackedContext::ListArgs::kBias)
+          .toTensor();
+  const float eps = api::utils::safe_downcast<float>(
+      layernorm_context->get_val(LayernormPackedContext::ListArgs::kEps)
+          .toDouble());
+
+  // We invoke native_layer_norm which returns a tuple of tensors: <layer_norm,
+  // mean, 1/sqrt(var+eps)>, but we only need the first tensor (layer_norm).
+  std::tuple<Tensor, Tensor, Tensor> native_layer_norm_output =
+      at::native_layer_norm(input, normalized_shape, weight_opt, bias_opt, eps);
+  return std::get<0>(native_layer_norm_output);
+}
 
 Tensor layer_norm(
     const at::Tensor& input_arg,
@@ -22,13 +85,11 @@ Tensor layer_norm(
     const c10::optional<Tensor>& bias_opt /* optional */,
     double eps,
     bool /* cudnn_enable, deprecated */) {
-  const Tensor input = input_arg.is_vulkan() ? input_arg : input_arg.vulkan();
-
-  // We invoke native_layer_norm which returns a tuple of tensors: <layer_norm,
-  // mean, 1/sqrt(var+eps)>, but we only need the first tensor (layer_norm).
-  std::tuple<Tensor, Tensor, Tensor> native_layer_norm_output =
-      at::native_layer_norm(input, normalized_shape, weight_opt, bias_opt, eps);
-  return std::get<0>(native_layer_norm_output);
+  return run_layernorm_context(
+      input_arg,
+      normalized_shape,
+      c10::make_intrusive<LayernormPackedContext>(
+          LayernormPackedContext(weight_opt, bias_opt, eps)));
 }
 
 #ifdef USE_VULKAN_API
@@ -39,7 +100,6 @@ TORCH_LIBRARY_IMPL(aten, Vulkan, m) {
 
 #endif /* USE_VULKAN_API */
 
-} // namespace
 } // namespace ops
 } // namespace vulkan
 } // namespace native