--- a/aten/src/ATen/native/vulkan/ops/Register.h
+++ b/aten/src/ATen/native/vulkan/ops/Register.h
@@ -7,6 +7,7 @@ namespace ops {
 
 int register_vulkan_conv2d_packed_context();
 int register_vulkan_linear_packed_context();
+int register_vulkan_layernorm_packed_context();
 
 } // namespace ops
 } // namespace vulkan